Rank Analysis/James Ranking functions/requirements copy.txt
pandas>=0.1
requests>=0.1
flask>=0.2
google-cloud-datastore>=0.1